Refactor git flowgraph component to be extensible
